Project_012/services/governance_gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
regex==2024.5.15
pyyaml==6.0.2
python-multipart>=0.0.7
spacy
transformers
requests
mlflow==3.1.1
prometheus_client

